Add working e2e tests with fixed issues

+++ python/rhythm/config.py
@@ -14,7 +14,7 @@ class Settings(BaseSettings):
     # Worker settings
     worker_heartbeat_interval: int = 5  # seconds
     worker_heartbeat_timeout: int = 30  # seconds
-    worker_poll_interval: int = 1  # seconds
+    worker_poll_interval: float = 1.0  # seconds (can be fractional for fast polling in tests)
     worker_max_concurrent: int = 10  # max concurrent executions per worker
 
     # Execution defaults
